Various fixes & additional convenience method regarding TS3 perms
